+++ EveryTime/Controller/Main/MainViewController+UICollectionView.swift
@@ -10,6 +10,25 @@ import UIKit
 
 extension MainViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
     
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: CollectionCellIds.UpgradeHeader.rawValue, for: indexPath) as! UpgradeHeaderView
+        header.theme = viewModel?.theme
+        header.setupView()
+        return header
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
+        guard let viewModel = viewModel else {
+            return CGSize(width: view.bounds.width, height: view.bounds.height / 3.5)
+        }
+        //boolean to check if user has pro
+        if (viewModel.isPro) {
+            return CGSize(width: view.bounds.width, height: 0.0)
+        } else {
+            return CGSize(width: view.bounds.width, height: view.bounds.height / 3.5)
+        }
+    }
+    
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         let cell = collectionView.cellForItem(at: indexPath) as! MainViewCell
         cell.animateCellForSelection()
@@ -21,11 +40,10 @@ extension MainViewController: UICollectionViewDataSource, UICollectionViewDelega
                 //        horizontalDelegate.dismissInteractor  = HorizontalTransitionInteractor(viewController: vc) // to be worked on. issue with timer when swiping to dismiss
                 vc.transitioningDelegate = horizontalDelegate
                 vc.modalPresentationStyle = .custom
-                
                 stopTimer()
                 self.present(vc, animated: true, completion: nil)
             }
-
+            
         }
     }